refactor(bread crumbs): break up different parts
