+++ src/components/BreadcrumbLink/BreadcrumbLink.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { ClaimType, GradeType, SubjectType } from './BreadcrumbModel';
+import { ClaimType, GradeType, SubjectType } from '../Breadcrumbs/BreadcrumbModel';
 import { Colors } from '../../constants';
 
 /**
